+++ src/main/java/net/soulsweaponry/mixin/KeyBindingAccessor.java
@@ -1,0 +1,13 @@
+package net.soulsweaponry.mixin;
+
+import net.minecraft.client.option.KeyBinding;
+import net.minecraft.client.util.InputUtil;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(KeyBinding.class)
+public interface KeyBindingAccessor {
+
+    @Accessor("boundKey")
+    InputUtil.Key getBoundKey();
+}