+++ src/main/java/net/soulsweaponry/mixin/ShieldItemMixin.java
@@ -7,6 +7,7 @@ import net.minecraft.item.ShieldItem;
 import net.minecraft.text.Text;
 import net.minecraft.world.World;
 import net.soulsweaponry.config.ConfigConstructor;
+import net.soulsweaponry.items.ITooltipInfo;
 import net.soulsweaponry.util.TooltipAbilities;
 import net.soulsweaponry.util.TooltipUtil;
 import org.jetbrains.annotations.Nullable;
@@ -26,7 +27,7 @@ public class ShieldItemMixin {
             if (Screen.hasShiftDown()) {
                 TooltipUtil.addAbilityTooltip(TooltipAbilities.PARRY, stack, tooltip);
             } else {
-                tooltip.add(Text.translatable("tooltip.soulsweapons.shift"));
+                ITooltipInfo.addShowInfoText(tooltip);
             }
         }
     }